Split responsibilities for call management and callstr construction between two classes
The base class, CommandLineCaller, doesn't concern itself with the callstr, other than to
receive it and pass it on to Popen. Really, this is the core of cliceo functionality. Facilites
for callstr construction, now moved out to child class SimpleGenericCLIcontroller, are
somewhere between a simple utility and a usage example.